+++ src/contracts/moblib.sol
@@ -2,11 +2,11 @@ pragma solidity ^0.4.16;
 
 library MobLib {
 
-    function max(int8 a, int8 b) public pure returns (int8) {
+    function max(int16 a, int16 b) public pure returns (int16) {
        return a > b ? a : b;
     }
 
-    function min(int8 a, int8 b) public pure returns (int8) {
+    function min(int16 a, int16 b) public pure returns (int16) {
        return a < b ? a : b;
     }
 }
